+++ history.md
@@ -1,3 +1,5 @@
+* Wall terrain types can be marked to not accept wall decor
+* Added extra dialogue to Bear Bastard's Mecha Camp for Criminal PCs
 * PlacableThing's altitude property now does something
 * Combat can be set to keep going even if there are no active enemies on map
 * Room contents with fixed positions will now be placed on map